Revert "`imports FOO.bar` (#2144)"
This reverts commit 62671d4bf6194a4d3b0411cf55a8abd0c8d63aec.
