--- k-distribution/tests/regression-new/import/Makefile
@@ -1,7 +1,0 @@
-DEF=test
-EXT=test
-TESTDIR=.
-KOMPILE_BACKEND=llvm
-KOMPILE_FLAGS=--syntax-module TEST
-
-include ../../../include/kframework/ktest.mak
