--- k-distribution/tests/regression-new/import/test.k
@@ -1,20 +1,0 @@
-module TEST
-  imports A
-
-  rule foo() => .K
-  rule alsoAFoo() => .K
-endmodule
-
-module A
-  imports B.foo
-endmodule
-
-module B
-  imports C
-  syntax Foo ::= foo()
-  syntax Bar ::= bar()
-endmodule
-
-module C
-  syntax Foo ::= alsoAFoo() [foo]
-endmodule
