+++ kernel/src/main/java/org/kframework/unparser/ToJson.java
@@ -48,13 +48,12 @@ import java.util.Optional;
 
 import javax.json.Json;
 import javax.json.stream.JsonGenerator;
+import javax.json.JsonWriterFactory;
 import javax.json.JsonArrayBuilder;
 import javax.json.JsonObject;
 import javax.json.JsonObjectBuilder;
-import javax.json.JsonStructure;
-import javax.json.JsonValue;
 import javax.json.JsonWriter;
-import javax.json.JsonWriterFactory;
+import javax.json.JsonStructure;
 
 import scala.Enumeration;
 import scala.Option;
@@ -144,14 +143,8 @@ public class ToJson {
         JsonArrayBuilder imports = Json.createArrayBuilder();
         stream(mod.imports()).forEach(i -> {
           JsonObjectBuilder jimp = Json.createObjectBuilder();
-          jimp.add("name", i.name());
-          jimp.add("isPublic", i.isPublic());
-          Optional<String> tag = mutable(i.tag()).map(t -> t.name());
-          if (tag.isPresent()) {
-            jimp.add("tag", tag.get());
-          } else {
-            jimp.add("tag", JsonValue.NULL);
-          }
+          jmod.add("name", i.name());
+          jmod.add("isPublic", i.isPublic());
           imports.add(jimp.build());
         });
